refactor: 요청에 주소가 없을 시 전체 기준 조회
